mcp_server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-server"
version = "0.1.0"
description = "Graphiti MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.5.0",
    "openai>=1.68.2",
    "graphiti-core>=0.8.2",
    "azure-identity>=1.21.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
py-modules = ["graphiti_mcp_server", "memory_enhanced_server", "memory_entity_types"]
